refactor: Se cambio la estructura de carpetas de recreo
